+++ src/main/java/com/letscareer/global/configuration/AppConfiguration.java
@@ -1,0 +1,9 @@
+package com.letscareer.global.configuration;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class AppConfiguration {
+}
