+++ src/main/java/com/letscareer/global/configuration/P6SpyFormatter.java
@@ -1,4 +1,4 @@
-package com.letscareer.global.config;
+package com.letscareer.global.configuration;
 
 import com.p6spy.engine.logging.Category;
 import com.p6spy.engine.spy.P6SpyOptions;
